fix(move_coordinates.bash): use tmpfiles correctly
The previous implementaion did not correctly iterate over the file. Each iteration acted on a clean input, so changes were not cumulative. To address this, an initial (temporary) copy of the GTF is created. For each loop, a tmp file is used to hold the output of awk, then moved back to the copied GTF. Finally, the copied gTF is moved to OUT, ensuring there is no waste.

This is a prime location for speed ups. This is the best pure bash implementation I could sort out.