feat(rules): support CellRanger
Due to ongoing file formatting issues with STAR, we have reverted to Cellranger. This commits also add support for the necessary mkgtf and mkref calls required for a custom cellranger reference.

BREAKING CHANGE: Though the reversion does not impact apparent function significantly, the outputs and underlying analysis are changed enough to warrant this BREAKING.